Compatible rk3399_linux project 64 cross compiler

+++ src/AudioPlayer.cpp
@@ -175,7 +175,7 @@ void AudioPlayer::setMessageHandler(MessageHandler *handler)
 
 long long AudioPlayer::getDuration()
 {
-    long long gstDuration = 0;
+    gint64 gstDuration = 0;
     int duration = -1;
     
     if (m_playbin && gst_element_query_duration(m_playbin, GST_FORMAT_TIME, &gstDuration))
